信息科学与工程学院/机器视觉实践/pyproject.toml
[project]
name = "image_processing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.1.1",
    "opencv-python>=4.10.0.84",
    "pillow>=10.4.0",
    "scikit-image>=0.24.0",
    "simple-term-menu>=1.6.4",
]
